Cast to double the int for correct expression

+++ Solver/Main.cpp
@@ -775,9 +775,9 @@ const int Main::oneThreadSolve()
         if (num % 4 == 3) myConf.fixRestartType = dynamic_restart;
         if (num % 4 == 2) myConf.doCalcReach = false;
         //else myConf.fixRestartType = static_restart;
-        myConf.simpBurstSConf *= 1.0 + num;
-        myConf.simpStartMult *= 1.0 + 0.2*num;
-        myConf.simpStartMMult *= 1.0 + 0.2*num;
+        myConf.simpBurstSConf *= 1 + num;
+        myConf.simpStartMult *= 1.0 + 0.2*(double)num;
+        myConf.simpStartMMult *= 1.0 + 0.2*(double)num;
         if (num == numThreads-1) {
             //myConf.doVarElim = false;
             myConf.doPerformPreSimp = false;
